ospf6d: add foreach_lsdesc iterator
Add a generic iterator for the descriptors in an LSA, with
specializations for the descriptor type.

This introduces a callback mechanism to decouple the LSA type from the
function that operates on each of the descriptors.

It prepares for further specialization of the iterators for LSAs that
contain TLVs, while reusing the callback that operate on the descriptors
contained in the TLVs.

Signed-off-by: Andrew Cooks <acooks.at.bda@gmail.com>