tests: add a test for legacy-only LSAs
The purpose of this test is to check that ospf6d can operate on Legacy
LSAs, without originating or receiving any of the TLV-based LSAs of
RFC 8362.

It checks for the presence of the Legacy LSAs, and when that passes, it
checks for the absence of any E-LSAs.

"Absence" means:
'not detected during a short time interval following the stable
convergence'

This obviously does not prove that an E-LSA could ever exist,
but hopefully it shows that E-LSAs can be disabled.

Signed-off-by: Andrew Cooks <acooks.at.bda@gmail.com>